// src/shared/api/generated/models/AccountProfilePutRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CRM API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AccountProfilePutRequest
 */
export interface AccountProfilePutRequest {
    /**
     * 
     * @type {string}
     * @memberof AccountProfilePutRequest
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof AccountProfilePutRequest
     */
    lastName: string;
    /**
     * Дата рождения пользователя (от 16 до 100 лет)
     * @type {Date}
     * @memberof AccountProfilePutRequest
     */
    birthDay: Date;
    /**
     * 
     * @type {boolean}
     * @memberof AccountProfilePutRequest
     */
    isPrivate: boolean;
}

/**
 * Check if a given object implements the AccountProfilePutRequest interface.
 */
export function instanceOfAccountProfilePutRequest(value: object): value is AccountProfilePutRequest {
    if (!('firstName' in value) || value['firstName'] === undefined) return false;
    if (!('lastName' in value) || value['lastName'] === undefined) return false;
    if (!('birthDay' in value) || value['birthDay'] === undefined) return false;
    if (!('isPrivate' in value) || value['isPrivate'] === undefined) return false;
    return true;
}

export function AccountProfilePutRequestFromJSON(json: any): AccountProfilePutRequest {
    return AccountProfilePutRequestFromJSONTyped(json, false);
}

export function AccountProfilePutRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccountProfilePutRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'firstName': json['first_name'],
        'lastName': json['last_name'],
        'birthDay': (new Date(json['birth_day'])),
        'isPrivate': json['is_private'],
    };
}

export function AccountProfilePutRequestToJSON(json: any): AccountProfilePutRequest {
    return AccountProfilePutRequestToJSONTyped(json, false);
}

export function AccountProfilePutRequestToJSONTyped(value?: AccountProfilePutRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'first_name': value['firstName'],
        'last_name': value['lastName'],
        'birth_day': ((value['birthDay']).toISOString().substring(0,10)),
        'is_private': value['isPrivate'],
    };
}